+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Upcoming
 
-- Nothing as of now
+- Dont only consider HD when calculating grade but also Flashlight and FadeIn
 
 # v0.1.0